task CROD fix

+++ app/Http/Controllers/TasksController.php
@@ -12,10 +12,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::orderBy('id','desc')->paginate(25);
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザーを取得
+            $user = \Auth::user();
+            $tasks = Task::where('user_id', $user->id)->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        return view('tasks.index', $data);
     }
 
     /**
@@ -23,10 +30,14 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+        if (\Auth::check()) { // 認証済みの場合
+            $task = new Task;
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -34,16 +45,19 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        // バリデーション
-        $request->validate([
-            'content' => 'required',
-            'status' => 'required|max:10',
-        ]);
-
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        if (\Auth::check()) { // 認証済みの場合
+            // バリデーション
+            $request->validate([
+                'content' => 'required',
+                'status' => 'required|max:10',
+            ]);
+            // 認証済みユーザー（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+            $request->user()->tasks()->create([
+                'content' => $request->content,
+                'status' => $request->status,
+            ]);
+            return redirect()->route('tasks.index')->with('success', 'タスクを作成しました。');
+        }
         return redirect('/');
     }
 
@@ -53,9 +67,11 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
     }
 
     /**
@@ -64,9 +80,14 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -74,14 +95,18 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'content' => 'required',
-            'status' => 'required|max:10',
-        ]);
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+            $request->validate([
+                'content' => 'required',
+                'status' => 'required|max:10',
+            ]);
+            $request->user()->tasks()->findOrFail($id)->update([
+                'content' => $request->content,
+                'status' => $request->status,
+            ]);
+            return redirect()->route('tasks.index')->with('success', 'タスクを更新しました。');
+        }
         return redirect('/');
     }
 
@@ -91,7 +116,14 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-        return redirect('/');
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success', 'Delete Successful');
+        }
+        // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 }
